requirements.txt
streamlit
tensorflow
pillow
numpy
pandas
openpyxl